pyproject.toml
[project]
name = "simple-datawarehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.12.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "tqdm>=4.67.1",
]
